This is a working version of the scraper and saver. Made some minor changes in Main.java and there is no bug when the final table is being scraped. Added some additional constant variables to Variables.java.
DBSaver.java is written with Apache.POI (Added with Maven, because I couldn't add libraries manually). This version of DBSaver.java is complete, for now it can create and save single spreadsheet.

Code itself is not yet ready, It needs some "cosmetics".

For the future I plan to add some functionality to both Main.java and DBSaver.java, which will make me able to scrape any desired region and save it with alongside other already-scraped regions.